+++ coding/ace/logtalk_highlight_rules.js
@@ -201,7 +201,8 @@ var LogtalkHighlightRules = function() {
          { token: 'punctuation.definition.string.begin.logtalk',
            regex: '"',
            push: 
-            [ { token: 'constant.character.escape.logtalk', regex: '\\\\.' },
+            [ { token: 'constant.character.escape.logtalk',
+                regex: '\\\\([\\\\abfnrtv"\']|(x[a-fA-F0-9]+|[0-7]+)\\\\)' },
               { token: 'punctuation.definition.string.end.logtalk',
                 regex: '"',
                 next: 'pop' },